Feat: 빈 input 체크 커스텀 훅 분리
